Updated README.md to direct users to get pyz files from releases, not the clones
As a first-time user, possibly not drinking enough coffee, I initially thought the sscollector.pyz file would be found in the GitHub repo clone that is the preferred method of installation for StormSpotter.  I've added a phrase and a link to direct people to the releases page to get this file.